+++ src/main/java/legend/game/combat/effects/MonsterDeathEffectObjectDestructor30.java
@@ -12,8 +12,8 @@ public class MonsterDeathEffectObjectDestructor30 {
 
   public int scaleModifier_04;
   public int scaleModifierVelocity_08;
-  public int angleModifier_0c;
-  public int angleModifierVelocity_10;
+  public float angleModifier_0c;
+  public float angleModifierVelocity_10;
   public final VECTOR translation_14 = new VECTOR();
 
   /** shorts */
@@ -23,7 +23,7 @@ public class MonsterDeathEffectObjectDestructor30 {
   public int stepR_2a;
   public int stepG_2c;
   public int stepB_2e;
-  
+
   public MonsterDeathEffectObjectDestructor30(final int index) {
     this.index = index;
   }
